Don't pack.

+++ convert.sh
@@ -46,7 +46,7 @@ all_chapter_dirs="installation \
 #   - summary: _summary.md
 echo "* Documents are all licensed under [Creative Commons](http://creativecommons.org/)." > ${INDEX_MD}
 echo "* Documents are all licensed under [Creative Commons](http://creativecommons.org/)." > ${README_MD}
-echo "* We're migrating [old wiki documents](http://www.iredmail.org/wiki) to Markdown for easier maintenance, converted documents are available [here](https://bitbucket.org/zhb/docs.iredmail.org/src). [Download all documents](iredmail-docs.tar.bz2) (HTML) for offline reading." >> ${INDEX_MD}
+echo "* We're migrating [old wiki documents](http://www.iredmail.org/wiki) to Markdown for easier maintenance, converted documents are available [here](https://bitbucket.org/zhb/docs.iredmail.org/src)." >> ${INDEX_MD}
 
 for chapter_dir in ${all_chapter_dirs}; do
     # Get articles
@@ -154,7 +154,9 @@ if echo "$@" | grep -q -- '--sync-local'; then
     # Copy to iredmail.com/docs/
     rm -rf /Volumes/STORAGE/Dropbox/Backup/iredmail.com/docs/*
     cp -rf html/* /Volumes/STORAGE/Dropbox/Backup/iredmail.com/docs/
-    cp -f ${PACK_TAR_NAME} /Volumes/STORAGE/Dropbox/Backup/iredmail.com/docs/
+    if [ -f ${PACK_TAR_NAME} ]; then
+        cp -f ${PACK_TAR_NAME} /Volumes/STORAGE/Dropbox/Backup/iredmail.com/docs/
+    fi
 
     rm -f ${PACK_TAR_NAME}
 fi
